+++ src/controllers/state.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { StateService } from "../services/state.service";
-import { HttpError } from "../utils/errors/HttpError";
+import { HttpError } from "../utils/httpError";
+import { isUUID } from "class-validator";
 
 const stateService = new StateService();
 
@@ -38,10 +39,17 @@ export class StateController {
 
   static async getById(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do estado não informado ou formato inválido",
+      });
+    }
+
     const state = await stateService.findById(id);
 
     if (!state) {
-      return res.status(404).json({ message: "State not found" });
+      return res.status(404).json({ message: "Estado não encontrado" });
     }
 
     return res.json(state);
@@ -49,10 +57,17 @@ export class StateController {
 
   static async update(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do estado não informado ou formato inválido",
+      });
+    }
+
     const updated = await stateService.update(id, req.body);
 
     if (!updated) {
-      return res.status(404).json({ message: "State not found" });
+      return res.status(404).json({ message: "Estado não encontrado" });
     }
 
     return res.json(updated);
@@ -60,10 +75,17 @@ export class StateController {
 
   static async delete(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do estado não informado ou formato inválido",
+      });
+    }
+
     const deleted = await stateService.delete(id);
 
     if (!deleted) {
-      return res.status(404).json({ message: "State not found" });
+      return res.status(404).json({ message: "Estado não encontrado" });
     }
 
     return res.status(204).send();
