+++ src/controllers/user.controller.ts
@@ -1,11 +1,12 @@
 import { Request, Response } from "express";
 import { UserService } from "../services/user.service";
+import { isUUID } from "class-validator";
 
 const userService = new UserService();
 
 export class UserController {
   static async create(req: Request, res: Response) {
-    const { name, email, password, profileType } = req.body;
+    const { name, email, password } = req.body;
 
     if (!name || name.trim() === "") {
       res.status(400).json({ message: "O campo 'name' é obrigatório" });
@@ -23,7 +24,6 @@ export class UserController {
       name,
       email,
       password,
-      profileType: "client",
     });
 
     return res.status(201).json(user);
@@ -36,10 +36,16 @@ export class UserController {
 
   static async getById(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do usuário não informado ou formato inválido",
+      });
+    }
     const user = await userService.findById(id);
 
     if (!user) {
-      return res.status(404).json({ message: "User not found" });
+      return res.status(404).json({ message: "Usuário não encontrado" });
     }
 
     return res.json(user);
@@ -47,10 +53,16 @@ export class UserController {
 
   static async update(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do usuário não informado ou formato inválido",
+      });
+    }
     const updated = await userService.update(id, req.body);
 
     if (!updated) {
-      return res.status(404).json({ message: "User not found" });
+      return res.status(404).json({ message: "Usuário não encontrado" });
     }
 
     return res.json(updated);
@@ -58,10 +70,16 @@ export class UserController {
 
   static async delete(req: Request, res: Response) {
     const { id } = req.params;
+
+    if (!id || isUUID(id)) {
+      return res.status(400).json({
+        message: "Id do usuário não informado ou formato inválido",
+      });
+    }
     const deleted = await userService.delete(id);
 
     if (!deleted) {
-      return res.status(404).json({ message: "User not found" });
+      return res.status(404).json({ message: "Usuário não encontrado" });
     }
 
     return res.status(204).send();
